Add script for recording the object sizes of mbed-client built on armcc
This script will collect the module size information of the mbed-client
and its current dependencies. Each module will get a separate txt-file
under results -directory.

Note: the numbers need to be taken with grain of salt, as the end result
of binary depends on the client, as if it does not call some paths, the
linker is free to remove them from the end binary. So, in practice these
numbers represent the worst case scenario.